requirements.txt
python-telegram-bot==13.15
requests==2.31.0
aiohttp==3.9.1
aiosqlite==0.19.0
python-dotenv==1.0.0
pytz
psycopg2-binary
colorama
aiofiles
psutil
bs4
